подправил тесты(все равно камни)

+++ src/main/java/com/statsystem/dbservice/DAO/ProjectDAO.java
@@ -17,7 +17,7 @@ import java.util.List;
 public interface ProjectDAO {
     void addProject(Project project) throws SQLException;
     void updateProject(Project project) throws SQLException;
-    Project getProjectById(Integer id) throws SQLException;
+    Project getProjectById(long id) throws SQLException;
     List<Project> getAllProjects() throws SQLException;
     void deleteProject(Project project) throws SQLException;
 }
